# Conflicts:
#	core/src/main/java/io/nuls/core/chain/entity/BaseNulsData.java
#	core/src/main/java/io/nuls/core/constant/ErrorCode.java

+++ account-impl/src/main/java/io/nuls/account/constant/AccountConstant.java
@@ -8,7 +8,7 @@ package io.nuls.account.constant;
  */
 public interface AccountConstant {
     //version
-    int ACCOUNT_MODULE_VERSION = 1111;
+    short ACCOUNT_MODULE_VERSION = 1111;
     //Minimum version supported
     int MINIMUM_VERSION_SUPPORTED = 0;
